+++ src/main/java/org/accounting/system/services/clients/Attributes.java
@@ -1,0 +1,9 @@
+package org.accounting.system.services.clients;
+
+import java.util.List;
+
+public class Attributes {
+
+    public List<String> description;
+    public List<String> defaultConfiguration;
+}