+++ src/main/java/org/accounting/system/services/clients/TokenResponse.java
@@ -1,0 +1,8 @@
+package org.accounting.system.services.clients;
+
+public class TokenResponse {
+
+    public String access_token;
+    public String token_type;
+    public int expires_in;
+}