+++ src/test/java/org/accounting/system/PrepareTest.java
@@ -1,5 +1,6 @@
 package org.accounting.system;
 
+import io.quarkus.test.InjectMock;
 import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.keycloak.client.KeycloakTestClient;
 import io.restassured.http.ContentType;
@@ -13,16 +14,20 @@ import org.accounting.system.repositories.metric.MetricRepository;
 import org.accounting.system.repositories.metricdefinition.MetricDefinitionRepository;
 import org.accounting.system.repositories.project.ProjectRepository;
 import org.accounting.system.repositories.provider.ProviderRepository;
+import org.accounting.system.services.KeycloakService;
+import org.accounting.system.services.clients.GroupMembership;
 import org.eclipse.microprofile.rest.client.inject.RestClient;
 import org.json.simple.parser.ParseException;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.TestInstance;
+import org.mockito.Mockito;
 
 import java.util.Set;
 import java.util.concurrent.ExecutionException;
 
 import static io.restassured.RestAssured.given;
+import static org.mockito.ArgumentMatchers.any;
 
 @QuarkusTest
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
@@ -44,6 +49,9 @@ public class PrepareTest {
     @Inject
     MetricRepository metricRepository;
 
+    @InjectMock
+    KeycloakService keycloakService;
+
     KeycloakTestClient keycloakClient = new KeycloakTestClient();
 
     @BeforeAll
@@ -58,6 +66,10 @@ public class PrepareTest {
 
     @BeforeEach
     public void before() throws ParseException {
+
+        Mockito.when(keycloakService.getValueByKey(any())).thenReturn("groupId");
+        Mockito.when(keycloakService.getConfiguration(any(), any())).thenReturn(new GroupMembership());
+
 
         metricDefinitionRepository.deleteAll();
         projectRepository.deleteAll();